% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lines.ccaf}
\alias{lines.ccaf}
\title{Add a plot a confidence curve of attenuated correlation coefficients.}
\usage{
\method{lines}{ccaf}(x, type = "l", col = "red3", lwd = 2, ...)
}
\arguments{
\item{x}{An object of class \code{ccaf}. The confidence curve to plot.}

\item{type}{The type of plot.}

\item{col}{The color of the curve.}

\item{lwd}{The thickness of the curve.}

\item{...}{Passed to \code{lines}.}
}
\value{
An invisible copy of \code{x}.
}
\description{
Add a plot a confidence curve of attenuated correlation coefficients.
}
