% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ccaf}
\alias{plot.ccaf}
\title{Plot a confidence curve of attenuated correlation coefficients.}
\usage{
\method{plot}{ccaf}(x, y, level = 0.95, ...)
}
\arguments{
\item{x}{An object of class \code{ccaf}. The confidence curve to plot.}

\item{y}{Ignored; supported for compatibility with the \code{plot} generic.}

\item{level}{Level to highlight. Defaults to 0.95. If \code{NULL},
highlights no level.}

\item{...}{Passed to \code{plot}.}
}
\value{
An invisible copy of \code{x}.
}
\description{
Plot a confidence curve of attenuated correlation coefficients.
}
