% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ccaf}
\alias{print.ccaf}
\title{Print method for \code{ccaf}}
\usage{
\method{print}{ccaf}(x, digits = 3, ...)
}
\arguments{
\item{x}{Object of class \code{ccaf}.}

\item{digits}{Passed to \code{signif}.}

\item{...}{Ignored.}
}
\value{
Invisible copy of \code{x}.
}
\description{
Print method for \code{ccaf}
}
