% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{p_value}
\alias{p_value}
\title{Calculate the p-value for an attenuated correlation coefficient.}
\usage{
p_value(rho, r, N)
}
\arguments{
\item{rho}{Numeric vector in [-1,1]. The correlation under the null
hypothesis.}

\item{r}{Numeric vector of three elemtents in [-1,1]. \code{r[1]} is the
correlation between the noisy measures X' and Y', \code{r[2]} is the
correlation between the noisy X' and the true X, while \code{r[3]} is
the correlation between the noisy Y' and the true Y.}

\item{N}{Numeric vector of three positive integers. \code{N[i]} is the
sample size for \code{r[i]}.}
}
\value{
Numeric in [0, 1]. The p-value under null-hypothesis rho.
}
\description{
Calculate the p-value for an attenuated correlation coefficient.
}
