% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{p_value}
\alias{p_value}
\title{Calculate the p-value for an attenuated correlation coefficient.}
\usage{
p_value(rho, r, N, positive = FALSE)
}
\arguments{
\item{rho}{Numeric vector in [-1,1]. The correlation under the null
hypothesis.}

\item{r}{Numeric vector of three elemtents in [-1,1]. \code{r[1]} is the
correlation between the noisy measures X' and Y', \code{r[2]} is the
correlation between the noisy X' and the true X, while \code{r[3]} is
the correlation between the noisy Y' and the true Y.}

\item{N}{Numeric vector of three positive integers. \code{N[i]} is the
sample size for \code{r[i]}.}

\item{positive}{Logical; If \code{TRUE}, the correlations \code{r[2]} and
\code{r[3]} are forced to be positive.}
}
\value{
Numeric in [0, 1]. The p-value under the null-hypothesis that the
    true correlation is rho.
}
\description{
Calculate the p-value for an attenuated correlation coefficient.
}
\examples{
    r = c(0.20, sqrt(0.45), sqrt(0.55))
    N = c(100, 100, 100)
    p_value(rho = 0, r, N) # Tests rho = 0.
}
